% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{activate.NetCDF}
\alias{activate.NetCDF}
\alias{active}
\alias{activate}
\alias{active<-}
\alias{activate}
\alias{active.NetCDF}
\alias{activate}
\alias{active<-.NetCDF}
\title{Activate}
\usage{
\method{activate}{NetCDF}(.data, what)

\method{active}{NetCDF}(x)

\method{active}{NetCDF}(x) <- value
}
\arguments{
\item{.data}{NetCDF object}

\item{what}{name of a variable}

\item{x}{NetCDF object}

\item{value}{name of variable to be active}
}
\value{
NetCDF object
}
\description{
Set the context for subsequent manipulations.
}
\details{
\code{activate} puts the named variable first
\code{active} gets and sets the active variable
}
\examples{
rnc <- NetCDF(system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package= "ncdump"))
activate(rnc, "palette")
}
