% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{dim_names}
\alias{dim_names}
\alias{dim_names}
\alias{dim_names.character}
\alias{dim_names}
\alias{dim_names.NetCDF}
\title{Dimension names}
\usage{
dim_names(x, ...)

\method{dim_names}{character}(x, ...)

\method{dim_names}{NetCDF}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
names of available dimensions
}
\description{
Dimension names
}
\examples{
example(NetCDF)
dim_names(rnc)
}
