% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{dimension_values}
\alias{dimension_values}
\alias{dimension_values.NetCDF}
\title{Dimension values}
\usage{
dimension_values(x)

\method{dimension_values}{NetCDF}(x)
}
\arguments{
\item{x}{}
}
\description{
The dimension values are the coordinates, the positions along each axis.
}
