% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{filtrate}
\alias{filtrate}
\title{Array subset by nse}
\usage{
filtrate(x, ...)
}
\arguments{
\item{x}{NetCDF object}

\item{...}{currently ignored}
}
\value{
data frame
}
\description{
NSE arguments must be named as per the dimensions in the variable. This is a restrictive variant of \code{dplyr::filter},
with a syntax more like \code{dplyr::mutate}. This ensures that each element is named, so we know which dimension to
apply this to, but also that the expression evaluated against can do some extra work for a nuanced test.
}
\examples{
## inst/dev/filtrate_early.R
##http://rpubs.com/cyclemumner/tidyslab
# 
}
