% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{hyper_index}
\alias{hyper_index}
\alias{hyper_index}
\alias{hyper_index.tbl_df}
\alias{hyper_index}
\alias{hyper_index.NetCDF}
\title{hyper slab index}
\usage{
hyper_index(x, ...)

\method{hyper_index}{tbl_df}(x, ...)

\method{hyper_index}{NetCDF}(x, ...)
}
\arguments{
\item{...}{expressions to \code{filtrate}}
}
\description{
hyper slab index
}
