% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{hyper_slice}
\alias{hyper_slice}
\alias{hyper_slice}
\alias{hyper_slice.hyperslab}
\alias{hyper_slice}
\alias{hyper_slice.NetCDF}
\alias{hyper_slice}
\alias{hyper_slice.character}
\title{hyper slice}
\usage{
hyper_slice(x, ...)

\method{hyper_slice}{hyperslab}(x, ...)

\method{hyper_slice}{NetCDF}(x, ...)

\method{hyper_slice}{character}(x, ...)
}
\arguments{
\item{x}{a hyper slab}

\item{...}{ignored}
}
\description{
hyper slice
}
