% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF.r
\name{ncatts}
\alias{ncatts}
\title{importFrom ncdf4 nc_open nc_close
.varnames <- function(x) {
names(.ndims(x))
}
.ndims <- function(x) {
nc <- nc_open(x)
dims <- sapply(nc$var, "[[", "ndims")
nc_close(nc)
dims
}}
\usage{
ncatts(x)
}
\arguments{
\item{x}{file path}
}
\description{
importFrom ncdf4 nc_open nc_close
.varnames <- function(x) {
names(.ndims(x))
}
.ndims <- function(x) {
nc <- nc_open(x)
dims <- sapply(nc$var, "[[", "ndims")
nc_close(nc)
dims
}
}
\keyword{internal}
