% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdump-package.R
\docType{package}
\name{ncdump}
\alias{ncdump-package}
\alias{ncdump}
\title{ncdump for NetCDF metadata.}
\description{
Provides a useable data structure of NetCDF metadata
NetCDF is Network Common Data Form \url{https://www.unidata.ucar.edu/software/netcdf/}.
}
\details{
There is one main function, returning an object of class "ncdump":
\tabular{ll}{
\code{\link{NetCDF}} \tab produces a list of data frames containing metadata from a NetCDF file\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/ncdump}
  \item Report bugs at \url{https://github.com/hypertidy/ncdump/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
