% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{nctivate}
\alias{nctivate}
\alias{nctivate.NetCDF}
\alias{nctive}
\title{Activate variable}
\usage{
nctivate(.data, what)

\method{nctivate}{NetCDF}(.data, what)

nctive(x)
}
\arguments{
\item{.data}{}

\item{what}{}
}
\description{
Activate variable
}
