% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF.r
\name{NetCDF}
\alias{NetCDF}
\title{Information about a NetCDF file, in convenient form.}
\usage{
NetCDF(x)
}
\arguments{
\item{x}{path to NetCDF file}
}
\description{
Information about a NetCDF file, in convenient form.
}

