% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{var_names}
\alias{var_names}
\alias{var_names}
\alias{var_names.character}
\alias{var_names}
\alias{var_names.NetCDF}
\title{Variable names}
\usage{
var_names(x, ...)

\method{var_names}{character}(x, ...)

\method{var_names}{NetCDF}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
names of variables available
}
\description{
Variable names
}
\examples{
example(NetCDF)
var_names(rnc)
}
