% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{varnames}
\alias{varnames}
\alias{varnames}
\alias{varnames.character}
\alias{varnames}
\alias{varnames.NetCDF}
\title{Variable names}
\usage{
varnames(x, ...)

\method{varnames}{character}(x, ...)

\method{varnames}{NetCDF}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
Variable names
}
