% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{activate}
\alias{activate}
\alias{activate.NetCDF}
\alias{nctive}
\title{Activate variable}
\usage{
activate(.data, what)

\method{activate}{NetCDF}(.data, what)

nctive(x)
}
\arguments{
\item{.data}{NetCDF object}

\item{what}{name of a variable}

\item{x}{NetCDF object}
}
\value{
NetCDF object
}
\description{
Stolen from tidygraph, we need this
}
\details{
\code{activate} puts the named variable first
\code{nctive} gets and sets the active variable
}
\examples{
example(NetCDF)
activate(rnc, "palette")
}
