% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{activate}
\alias{activate}
\alias{activate.NetCDF}
\alias{nctive}
\alias{print.NetCDF}
\title{Activate variable}
\usage{
activate(.data, what)

\method{activate}{NetCDF}(.data, what)

nctive(x)

\method{print}{NetCDF}(x)
}
\arguments{
\item{.data}{}

\item{what}{}
}
\description{
Stolen from tidygraph, we need this
}
\examples{
f <- "/rdsi/PRIVATE/raad/data/eclipse.ncdc.noaa.gov/pub/OI-daily-v2/NetCDF/2017/AVHRR/avhrr-only-v2.20170502_preliminary.nc"
x <- NetCDF(f)
nctive(x)
x <- activate(x, "sst") 
}
