% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{filtrate}
\alias{filtrate}
\title{Array subset by nse}
\usage{
filtrate(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
NSE arguments must be named as per the dimensions in the variable. This is a restrictive variant of \code{dplyr::filter},
with a syntax more like \code{dplyr::mutate}. This ensures that each element is named, so we know which dimension to
apply this to, but also that the expression evaluated against can do some extra work for a nuanced test.
}
\examples{
library(ncdump)
f <- "/rdsi/PRIVATE/raad/data/eclipse.ncdc.noaa.gov/pub/OI-daily-v2/NetCDF/2017/AVHRR/avhrr-only-v2.20170502_preliminary.nc"
x <- NetCDF(f)
x
nctive(x)
## push sst to the front for an extraction
x <- activate(x, "sst")
hyper_slab <- filtrate(x, lon = lon > 100, lat = lat < 30)
## it's alive, test the extraction
library(ncdf4)
nc <- nc_open(f)
library(dplyr)
var <- ncvar_get(nc, "sst", start = bind_rows(hyper_slab)$start, count = bind_rows(hyper_slab)$count)
image(var)
## push a different var to the front
x <- activate(x, "anom")
hyper_slab <- filtrate(x, lon = between(lon, 147, 250), lat = between(lat, -42, 20))
var <- ncvar_get(nc, "sst", start = bind_rows(hyper_slab)$start, count = bind_rows(hyper_slab)$count)
image(var)

# 
}
