% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{filtrate}
\alias{filtrate}
\title{Array subset by nse}
\usage{
filtrate(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
NSE arguments must be named as per the dimensions in the variable. This is a restrictive variant of \code{dplyr::filter},
with a syntax more like \code{dplyr::mutate}. This ensures that each element is named, so we know which dimension to
apply this to, but also that the expression evaluated against can do some extra work for a nuanced test.
}
\examples{
f <- "/rdsi/PRIVATE/raad/data/eclipse.ncdc.noaa.gov/pub/OI-daily-v2/NetCDF/2017/AVHRR/avhrr-only-v2.20170502_preliminary.nc"
x <- NetCDF(f)
nctive(x)
x <- nctivate(x, "sst")
filtrate(x, lon = lon > 100, lat = lat < 30)
}
