% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{nctivate}
\alias{nctivate}
\alias{nctivate.NetCDF}
\alias{nctive}
\alias{print.NetCDF}
\title{Activate variable}
\usage{
nctivate(.data, what)

\method{nctivate}{NetCDF}(.data, what)

nctive(x)

\method{print}{NetCDF}(x)
}
\arguments{
\item{.data}{}

\item{what}{}
}
\description{
Activate variable
}
\examples{
f <- "/rdsi/PRIVATE/raad/data/eclipse.ncdc.noaa.gov/pub/OI-daily-v2/NetCDF/2017/AVHRR/avhrr-only-v2.20170502_preliminary.nc"
x <- NetCDF(f)
nctive(x)
x <- nctivate(x, "sst") 
}
