% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF.r
\name{NetCDF}
\alias{NetCDF}
\title{Information about a NetCDF file, in convenient form.}
\usage{
NetCDF(x)
}
\arguments{
\item{x}{path to NetCDF file}
}
\description{
Information about a NetCDF file, in convenient form.
}
\examples{
rnc <- NetCDF(system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package= "ncdump"))
rnc
}
