% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF.r
\name{vars}
\alias{atts}
\alias{atts.NetCDF}
\alias{dims}
\alias{dims.NetCDF}
\alias{vars}
\alias{vars.NetCDF}
\title{NetCDF file description functions.}
\usage{
vars(x, ...)

\method{vars}{NetCDF}(x, ...)

dims(x, ...)

\method{dims}{NetCDF}(x, ...)

atts(x, ...)

\method{atts}{NetCDF}(x, ...)
}
\description{
NetCDF file description functions.
}

