% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF.r
\name{vars}
\alias{vars}
\alias{vars.NetCDF}
\alias{dims}
\alias{dims.NetCDF}
\alias{dimvars}
\alias{dimvars.NetCDF}
\alias{atts}
\alias{atts.NetCDF}
\title{NetCDF file description functions.}
\usage{
vars(x, ...)

\method{vars}{NetCDF}(x, ...)

dims(x, ...)

\method{dims}{NetCDF}(x, ...)

dimvars(x, ...)

\method{dimvars}{NetCDF}(x, ...)

atts(x, ...)

\method{atts}{NetCDF}(x, varname = "globalatts", ...)
}
\arguments{
\item{x}{NetCDF metadata object}

\item{...}{ignored}

\item{varname}{name of variable to get atts of (not yet implemented)}
}
\description{
NetCDF file description functions.
}
